requirements.txt
alpaca-py>=0.42.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
ta>=0.10.0
SQLAlchemy>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
APScheduler>=3.10.0
httpx>=0.24.0
openai>=1.0.0
PyYAML>=6.0.0
plotly>=5.15.0
websockets>=12.0,<15.0
lightgbm>=4.0.0
scikit-learn>=1.2.0
joblib>=1.3.0
